requirements.txt
apache-airflow==2.7.1
causalimpact==0.2.6
matplotlib==3.7.1
numpy==1.25.2
orbit==0.2
pandahouse==0.2.7
pandas==2.0.3
python-dotenv==1.0.0
python-telegram-bot==20.6
scipy==1.11.3
seaborn==0.13.0
statsmodels==0.14.0
swifter==1.4.0
tensorflow==2.14.0
tensorflow_probability==0.22.0
